DBT Orientation as multiple tiddlers
So I had my tiddlywiki saving browser extension break on me & had to use the github saver to avoid losing everything. But I broke up the DBT Orientation into multiple tiddlers to make it easier to have an overview (the list the leads to the tiddlers) to help me getting a better idea of it as a whole, and so I can expand on each of them without turning the orientation into this super long tiddler that just goes on and on. At least for now, some of these new tiddlers are stupid short; at some point maybe I'll have more info to add to them / write some content for them. I am starting to deviate a bit from the DBT skills manual, because honestly there are times where the way it is organized doesn't make sense to me, and this site is for me so I'm going to put things in ways that make sense to me.