Manual Alt's
Added some alternatives to Marsh'a DBT Skills Manual, since the thing is so so so hard to read and I think I've found some solid "for clients" types of resources. I'm not totally sure, but I'm putting them in here anyway. If I find that I hate them, i will later remove them.